refactor hud_income_level.py and create a folder for hud_income_level_factor
