+++ policyengine_us/tests/policy/baseline/gov/local/co/denver/dhs/property_tax_relief/co_denver_integration.yaml
@@ -8,7 +8,6 @@
     has_co_denver_dhs_elderly_disabled: true
     tax_unit_dependents: 0 
     in_denver: true
-    hud_income_level: MODERATE
   output:
     co_denver_property_tax_relief_income: 52_000
     co_denver_property_tax_relief_homeowner_eligible: true
@@ -27,7 +26,6 @@
     has_co_denver_dhs_elderly_disabled: true
     tax_unit_dependents: 0 
     in_denver: true
-    hud_income_level: MODERATE
   output:
     co_denver_property_tax_relief_income: 22_000
     co_denver_property_tax_relief_homeowner_eligible: false
@@ -54,7 +52,6 @@
       spm_unit:
         members: [person1, person2]
         has_co_denver_dhs_elderly_disabled: true
-        hud_income_level: MODERATE
     tax_units:
       tax_unit:
         members: [person1, person2]
