+++ policyengine_us/tests/policy/baseline/gov/local/co/denver/dhs/property_tax_relief/co_denver_property_tax_relief_homeowner_eligible.yaml
@@ -6,7 +6,6 @@
     real_estate_taxes: 1_000
     has_co_denver_dhs_elderly_disabled: true
     tax_unit_dependents: 0 
-    hud_income_level: MODERATE
     in_denver: true
   output:
     co_denver_property_tax_relief_homeowner_eligible: false 
@@ -20,7 +19,6 @@
     real_estate_taxes: 1_000
     has_co_denver_dhs_elderly_disabled: true
     tax_unit_dependents: 0 
-    hud_income_level: MODERATE
     in_denver: true
   output:
     co_denver_property_tax_relief_homeowner_eligible: true
@@ -33,7 +31,6 @@
     real_estate_taxes: 1_000
     has_co_denver_dhs_elderly_disabled: false
     tax_unit_dependents: 0 
-    hud_income_level: MODERATE
     in_denver: true
   output:
     co_denver_property_tax_relief_homeowner_eligible: false
@@ -46,7 +43,6 @@
     real_estate_taxes: 1_000
     has_co_denver_dhs_elderly_disabled: true
     tax_unit_dependents: 0 
-    hud_income_level: MODERATE
     in_denver: true
   output:
     co_denver_property_tax_relief_homeowner_eligible: true
@@ -59,7 +55,6 @@
     real_estate_taxes: 1_000
     has_co_denver_dhs_elderly_disabled: true
     tax_unit_dependents: 0 
-    hud_income_level: MODERATE
     in_denver: true
   output:
     co_denver_property_tax_relief_homeowner_eligible: true
