+++ policyengine_us/tests/policy/baseline/gov/local/co/denver/dhs/property_tax_relief/co_denver_property_tax_relief_renter_eligible.yaml
@@ -6,7 +6,6 @@
     rent: 5_000
     has_co_denver_dhs_elderly_disabled: true
     in_denver: true
-    hud_income_level: MODERATE
   output:
     co_denver_property_tax_relief_renter_eligible: false  
 
@@ -18,7 +17,6 @@
     rent: 1_000
     has_co_denver_dhs_elderly_disabled: true
     in_denver: true
-    hud_income_level: MODERATE
   output:
     co_denver_property_tax_relief_renter_eligible: true
 
@@ -30,7 +28,6 @@
     rent: 1_000
     has_co_denver_dhs_elderly_disabled: true 
     in_denver: true
-    hud_income_level: MODERATE
   output:
     co_denver_property_tax_relief_renter_eligible: true
 
@@ -42,7 +39,6 @@
     rent: 1_000
     has_co_denver_dhs_elderly_disabled: false
     in_denver: true
-    hud_income_level: MODERATE
   output:
     co_denver_property_tax_relief_renter_eligible: false
 
@@ -54,6 +50,5 @@
     rent: 1_000
     has_co_denver_dhs_elderly_disabled: true 
     in_denver: true
-    hud_income_level: MODERATE
   output:
     co_denver_property_tax_relief_renter_eligible: true
